% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbinFullRegular.R
\name{hexbinFullRegular}
\alias{hexbinFullRegular}
\title{Title}
\usage{
hexbinFullRegular(
  x,
  y = NULL,
  xbins = 128,
  xbnds = range(x),
  ybnds = range(y),
  xlab = NULL,
  ylab = NULL,
  IDs = FALSE
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{xbins}{}

\item{xbnds}{}

\item{ybnds}{}

\item{xlab}{}

\item{ylab}{}

\item{IDs}{}
}
\description{
Title
}
