% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexContour.R
\name{hexContour}
\alias{hexContour}
\title{Generate contour using a hexagonal grid.}
\usage{
hexContour(hexDensity, levels)
}
\arguments{
\item{hexDensity}{hexDensity object to be contoured.}

\item{levels}{Numeric vector for which contour lines should be generated}
}
\value{
list of x, y, and ID, for the lines contouring each levels.
}
\description{
This is made to follow the same output format as the isoband package. A dedicated plotting function is in the work. In the mean time, see example of how to plot the output with ggplot2::geom_path
Algorithm is a modification of the meandering triangles as described here https://blog.bruce-hill.com/meandering-triangles
}
\examples{

set.seed(133)
x=rnorm(200)
y=rnorm(200)
d = hexDensity(x=x,y=y,bandwidth=0.4)
cutoff=quantile(d@count,0.9)
lines = hexContour(d,cutoff)

library(ggplot2)

#plot against density
ggplot()+
  geom_point(
    aes(x=hcell2xy(d)$x,
        y=hcell2xy(d)$y,
        col=d@count)
 ) +
  scale_color_viridis_c()+
  geom_path(
    aes(
     x = lines[[1]]$x, y = lines[[1]]$y, group = lines[[1]]$id
    )
  )

#plot against data points
ggplot() +
  geom_point(
    aes(x=x,y=y)) +
  geom_path(
    aes(
      x = lines[[1]]$x, y = lines[[1]]$y, group = lines[[1]]$id
    )
  )
}
