% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexDensity.R
\name{hexDensity}
\alias{hexDensity}
\alias{hexDensity.default}
\title{Kernel Density Estimation with Hexagonal binning.}
\usage{
hexDensity(x, ...)

\method{hexDensity}{default}(
  x,
  y = NULL,
  xbins = 128,
  sigma = 1,
  edge = TRUE,
  diggle = FALSE,
  weight = NULL
)
}
\arguments{
\item{x, y}{Coords of the points or a single plotting structure to be used in binning. See xy.coords.}

\item{xbins}{Number of bins in a row.}

\item{sigma}{Bandwidth for kernel density calculation.}

\item{edge}{Logical value for whether to apply edge correction. Default is TRUE.}

\item{diggle}{Logical value for apply edge correction with the more accurate Jones-Diggle methor (need 'edge' to be TRUE).}

\item{weight}{numeric weight vector to be assigned to points.}
}
\value{
hexbin object.
}
\description{
Kernel Density Estimation with Hexagonal binning.
}
\examples{

set.seed(133)
d = hexDensity(x=rnorm(200),y=rnorm(200),sigma=0.15)
}
