% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexDensity.R
\name{hexDensity}
\alias{hexDensity}
\title{Title}
\usage{
hexDensity(x, y = NULL, xbins = 128, sigma = 1, edge = TRUE, diggle = FALSE)
}
\arguments{
\item{x}{}

\item{y}{}

\item{xbins}{}

\item{sigma}{}

\item{edge}{}

\item{diggle}{}
}
\description{
Title
}
