% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexDensity.R
\name{hexDensity}
\alias{hexDensity}
\alias{hexDensity.default}
\title{Kernel Density Estimation with Hexagonal binning.}
\usage{
hexDensity(x, ...)

\method{hexDensity}{default}(
  x,
  y = NULL,
  xbins = 128,
  bandwidth = NULL,
  edge = TRUE,
  diggle = FALSE,
  weight = NULL
)
}
\arguments{
\item{x, y}{Coords of the points or a single plotting structure to be used in binning. See xy.coords.}

\item{xbins}{Number of bins in a row.}

\item{bandwidth}{Bandwidth for the smoothing kernel. Either a single number or a vector of length 2 for the bandwidths in the x and y directions, respectively.}

\item{edge}{Logical value for whether to apply edge correction. Default is TRUE.}

\item{diggle}{Logical value for apply edge correction with the more accurate Jones-Diggle method (need 'edge' to be TRUE).}

\item{weight}{numeric weight vector to be assigned to points.}
}
\value{
hexbin object.
}
\description{
Default bandwidth is the diagonal plug-in bandwidth calculated by the package ks
}
\examples{

set.seed(133)
d = hexDensity(x=rnorm(200),y=rnorm(200),bandwidth=0.15)
}
