% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHexDensity.R
\name{plotHexDensity}
\alias{plotHexDensity}
\title{Title}
\usage{
plotHexDensity(
  kernel,
  colramp = colorRampPalette(viridis::viridis(11)),
  main = deparse(substitute(kernel)),
  legend = F,
  colorcut = seq(0, 1, length = 1024)
)
}
\arguments{
\item{kernel}{}

\item{colramp}{}

\item{main}{}

\item{legend}{}

\item{colorcut}{}
}
\description{
Title
}
