% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHexDensity.R
\name{plotHexDensity}
\alias{plotHexDensity}
\title{Plot method for hexDensity}
\usage{
plotHexDensity(
  hexDensity,
  colramp = colorRampPalette(viridis::viridis(11)),
  main = deparse(substitute(hexDensity)),
  legend = F,
  colorcut = seq(0, 1, length = 1024)
)
}
\arguments{
\item{hexDensity}{hexbin object returned by hexDensity}

\item{colramp}{Color function that accept an integer n and return n colors.}

\item{main}{Main title}

\item{legend}{Legend is currently non-functional and should be ignored.}

\item{colorcut}{Vector of values covering [0, 1] that determine hexagon color class boundaries and hexagon legend size boundaries. Alternatively, an integer (<= maxcnt) specifying the number of equispaced colorcut values in [0,1]. See hexbin.}
}
\description{
Adapted plotting function for hexbin object.
}
