% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2hcell.R
\name{xy2hcell}
\alias{xy2hcell}
\title{Get the hexagonal cell ID from a hexbin object at the specified xy coordinates.}
\usage{
xy2hcell(hexbin, x, y = NULL)
}
\arguments{
\item{hexbin}{hexbin object to be referenced to.}

\item{x, y}{coordinates or vectors of coordinates of the points.}
}
\value{
a vector the same length as x with the hexagonal cell ID for each point
}
\description{
Get the hexagonal cell ID from a hexbin object at the specified xy coordinates.
}
\examples{
set.seed(133)
d=hexbin(x=rnorm(20000),y=rnorm(20000),xbins=50)
xy2hcell(d,x=0.5,y=0.2)
}
